// src/doctor/infrastructure/repositories/orm-doctor-criterias.dto.ts
import { DoctorSpecialty } from "src/doctor/domain/value-objects/doctor-specialty.enum";

export interface OrmDoctorCriteriasDto {
    firstName?: string;
    middleName?: string;

    firstSurname?: string;
    secondSurname?: string;

    latitude?: number;
    longitude?: number;

    rating?: number;

    specialty?: DoctorSpecialty
}